refactor(components): разделить на компоненты форму работы с пользователями
